pyproject.toml
[project]
name = "dfm-server"
version = "0.2.0"
requires-python = ">=3.10"
dependencies = ["torch>=2.0", "fastapi", "uvicorn[standard]", "pydantic-settings", "rich", "requests", "matplotlib", "numpy", "transformers"]

[tool.setuptools.packages.find]
include = ["dfm_server*"]

[project.scripts]
dfm-server = "dfm_server.server:main"
